move music modal to settings tab
