+++ app/components/settings-modal/music-tab/SpotifyPlayer.tsx
@@ -10,7 +10,7 @@ export const SpotifyPlayer = () => {
           title="Spotify Playlist"
           loading="lazy"
           allow="autoplay; clipboard-write; encrypted-media 'self' https://open.spotify.com; fullscreen; picture-in-picture; "
-          sandbox="allow-scripts allow-same-origin allow-popups allow-forms"
+          sandbox="allow-scripts allow-same-origin"
         />
       </SpotifyEmbed>
     </MusicPlayer>
